feat: refactor MIAlert and StyledAlert components, move AllowedAlertSeverity type to a new file, and update imports
