+++ src/components/MIAlert/index.ts
@@ -1,3 +1,2 @@
 export * from './MIAlert';
 export * from './StyledAlert';
-export * from './types';
